% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_tools.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to database}
\usage{
connect_to_db()
}
\arguments{
\item{environment_file}{the environment variable file to execute}
}
\value{
a connector object to a Postgresql database
}
\description{
This function connects to the application database. It needs the database parameters to be entered as environment variables as OANALYSER_DB_HOST, OANALYSER_DB_PASS, OANALYSER_DB_USER, OANALYSER_DB_PORT.
}
\examples{
\dontrun{

con <- connect_to_db()
}


}
