% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_get_mistakes.R
\name{get_intercontrols}
\alias{get_intercontrols}
\title{Returns intercontrols parameters}
\usage{
get_intercontrols(number_of_controls = 20)
}
\arguments{
\item{number_of_controls}{}
}
\value{
the a character vector sarting with 'Start - 1'
}
\description{
Returns intercontrols parameters
}
\examples{
get_intercontrols(20)
}
