% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_get_mistakes.R
\name{get_unvariant_mistake_form}
\alias{get_unvariant_mistake_form}
\title{This function returns the unvariant mistake form}
\usage{
get_unvariant_mistake_form()
}
\description{
This function returns the unvariant mistake form
}
