% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_get_competition_params.R
\name{mod_get_competition_params}
\alias{mod_get_competition_params}
\title{Competition params server side}
\usage{
mod_get_competition_params(input, output, session)
}
\arguments{
\item{session}{}
}
\description{
Deals with the values of the comptetion parameters
}
