% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_get_competition_params.R
\name{mod_get_competition_paramsInput}
\alias{mod_get_competition_paramsInput}
\title{Competition params ui}
\usage{
mod_get_competition_paramsInput()
}
\value{
a form for inputing competition, course and performance parameters
}
\description{
Competition params ui
}
