% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_get_mistakes.R
\name{mod_get_mistakes}
\alias{mod_get_mistakes}
\title{Server side function for the get mistake form}
\usage{
mod_get_mistakes(input, output, session, number_of_controls = 20)
}
\arguments{
\item{session}{}
}
\description{
This server function helps adapting the lenght of the intercontrols radio group button
}
