% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_get_mistakes.R
\name{mod_get_mistakesInput}
\alias{mod_get_mistakesInput}
\title{Mistakes form}
\usage{
mod_get_mistakesInput()
}
\arguments{
\item{number_of_controls}{}
}
\value{
a mistake form
}
\description{
This function prepares the mistakes forms depending on the number of mistakes
}
