% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_get_mistakes.R
\name{mod_print_mistakes_committed}
\alias{mod_print_mistakes_committed}
\title{Print mistakes committed}
\usage{
mod_print_mistakes_committed()
}
\description{
This module prints the mistakes committed by the user dynamically as the user enters them.
}
\examples{
\dontrun{
mod_print_mistakes_commited
}

}
