% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_get_mistakes.R
\name{get_unvariant_mistake_form}
\alias{get_unvariant_mistake_form}
\title{This function returns the unvariant mistake form}
\usage{
get_unvariant_mistake_form()
}
\value{
a list of fluid rows objects
}
\description{
The unvariant mistake form is the type of mistake committed and the tiem lost on that particular mistake
}
