% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_get_competition_params.R
\name{mod_get_competition_params}
\alias{mod_get_competition_params}
\title{Competition params server side}
\usage{
mod_get_competition_params(input, output, session)
}
\arguments{
\item{input}{input dialog between server and ui}

\item{output}{output between serer and ui}

\item{session}{session object passed to server}
}
\value{
nothing
}
\description{
Deals with the values of the comptetion parameters
}
