% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_get_competition_params.R
\name{mod_get_competition_params}
\alias{mod_get_competition_params}
\title{Competition params server side}
\usage{
mod_get_competition_params(input, output, session, val_comp_param,
  input_val)
}
\arguments{
\item{input}{input dialog between server and ui}

\item{output}{output between serer and ui}

\item{session}{session object passed to server}
}
\value{
nothing
}
\description{
Registers the competition parameters each time the validate button is clicked on
}
